+++ tests/test_module.py
@@ -1,5 +1,5 @@
 import unittest
-from application import app
+from src.application import app
 
 
 class BasicTestCase(unittest.TestCase):
